Moves out the filter from the view file
